rutas con middle

+++ routes/web.php
@@ -22,12 +22,12 @@ Route::get('/', HomeController::class)->name('home');
 
 //INSTITUTES
 Route::get('institutes', [InstituteController::class, 'index'])->name('institutes.index');
-Route::get('institutes/create', [InstituteController::class, 'create'])->name('institutes.create');
-Route::post('institutes', [InstituteController::class, 'store'])->name('institutes.store');
+Route::get('institutes/create', [InstituteController::class, 'create'])->middleware(['auth','admin'])->name('institutes.create');
+Route::post('institutes', [InstituteController::class, 'store'])->middleware(['auth','admin'])->name('institutes.store');
 Route::get('institutes/{institute}', [InstituteController::class, 'show'])->name('institutes.show');
-Route::get('institutes/{institute}/edit', [InstituteController::class, 'edit'])->name('institutes.edit');
-Route::post('institutes/{institute}', [InstituteController::class, 'update'])->name('institutes.update');
-Route::delete('institute/{institute}', [InstituteController::class,'destroy'])->name('institutes.destroy');
+Route::get('institutes/{institute}/edit', [InstituteController::class, 'edit'])->middleware(['auth','admin'])->name('institutes.edit');
+Route::post('institutes/{institute}', [InstituteController::class, 'update'])->middleware(['auth','admin'])->name('institutes.update');
+Route::delete('institute/{institute}', [InstituteController::class,'destroy'])->middleware(['auth','admin'])->name('institutes.destroy');
 
 
 Route::get('/dashboard', function () {
